super-research: the route table is a module, transport is its one address
`transport.py` held 1,182 lines of which 442 were per-platform data:
`ROUTE_CONSTANTS`, `PUBLIC_CLIENT_CREDENTIALS`, and everything the two need
to stand alone. They move to `routes.py` — declarations only, importing no
sibling — and `transport.py` re-exports every name under the name it always
had, so all 42 `transport.ROUTE_CONSTANTS` references and all 27
`transport.PUBLIC_CLIENT_CREDENTIALS` references resolve unchanged and no call
site moves. The shape is `probes.py` + `smoke.py`: data in one module, the
mechanism that reads it in another. The seam keeps the opener, https-only,
method admission, the byte cap, the refusal parsing and the captive-portal
detector, and drops to 783 lines.

442 lines left rather than the 338 the two tables occupy. The route id block,
the credential id block, the placement vocabulary, `JSON_CONTENT_TYPE` and both
dataclasses are all named by the tables at construction time, and `routes.py`
can import nothing back. `REDDIT_SITE_ORIGIN` moves for a second reason:
leaving a reachable host behind would split the allowlist across two files,
and one auditable closed allowlist in one file is the binding constraint this
split has to keep. It keeps it — every origin, endpoint and credential the
package can name is now in `routes.py` and nowhere else.

This is a move, not an edit, and a green suite does not prove that. Checked by
value against `ba0d103`: all twelve `RouteConstant` fields across eighteen
routes and all five `PublicClientCredential` fields across three credentials
are identical, as is the sorted list of every origin, origin+path and
credential value. 437 of the 442 moved lines are byte-identical; the five that
are not are the four docstring lines the split falsifies and one comment whose
sentence was true of the seam and is not true of a data module.

`routes` joins `ROUTE_OWNING_MODULES` and deliberately not
`NETWORK_SEAM_MODULES` — owning an address is not permission to open a socket.
Falsified beside the tree: `import urllib.request` in `routes.py` reddens both
network scans. `test_adapters.py` carried a fifth `"transport.py"` filename
literal, scanning for the credential values that now live next door; it reads
the declaration now, like the four sites before it.
