+++ .orchflows/skills/super-research/tests/test_dependency_boundary.py
@@ -8,7 +8,7 @@ enumeration is shown to reject a module beside the tree that breaks it.
 
 Four things are enumerated, and only the first is transcribed by hand:
 
-*The module set.* The core's fourteen modules are spelled out, so a new sibling
+*The module set.* The core's fifteen modules are spelled out, so a new sibling
 joins by editing this file or not at all. The count is in the sentence and in
 `CORE_MODULES`, and a test below compares them: this docstring said eleven for
 three modules longer than it was true. The adapter modules are not spelled
@@ -80,6 +80,7 @@ CORE_MODULES = (
     "probes",
     "project",
     "router",
+    "routes",
     "runner",
     "schema",
     "smoke",
@@ -102,6 +103,7 @@ CORE_IMPORT_EDGES = {
     "probes": ("transport",),
     "project": ("schema",),
     "router": ("adapters", "schema"),
+    "routes": (),
     "runner": (
         "adapters",
         "cache",
@@ -115,7 +117,7 @@ CORE_IMPORT_EDGES = {
     ),
     "schema": (),
     "smoke": ("probes", "runner", "schema", "transport"),
-    "transport": (),
+    "transport": ("routes",),
 }
 
 # Everything the package takes from outside itself. Fifteen names, and the
